Extract review item component
